[REFACTOR] #54 - UiState통일 + BestViewModel 분리
