+++ app/src/main/java/co/kr/woowahan_banchan/presentation/ui/main/best/BestFragment.kt
@@ -3,6 +3,7 @@ package co.kr.woowahan_banchan.presentation.ui.main.best
 import android.os.Bundle
 import android.view.View
 import androidx.fragment.app.activityViewModels
+import androidx.fragment.app.viewModels
 import androidx.lifecycle.flowWithLifecycle
 import androidx.lifecycle.lifecycleScope
 import androidx.recyclerview.widget.LinearLayoutManager
@@ -12,16 +13,19 @@ import co.kr.woowahan_banchan.presentation.adapter.BestItemAdapter
 import co.kr.woowahan_banchan.presentation.ui.base.BaseFragment
 import co.kr.woowahan_banchan.presentation.ui.productdetail.ProductDetailActivity
 import co.kr.woowahan_banchan.presentation.viewmodel.MainViewModel
-import co.kr.woowahan_banchan.presentation.viewmodel.MainViewModel.UiState
+import co.kr.woowahan_banchan.presentation.viewmodel.UiState
+import co.kr.woowahan_banchan.presentation.viewmodel.main.BestViewModel
+import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.onEach
 
+@AndroidEntryPoint
 class BestFragment : BaseFragment<FragmentBestBinding>() {
 
     override val layoutRes: Int
         get() = R.layout.fragment_best
 
-    private val activityViewModel by activityViewModels<MainViewModel>()
+    private val viewModel by viewModels<BestViewModel>()
     private val bestAdapter = BestItemAdapter { title, hash -> startDetailActivity(title, hash) }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
@@ -34,10 +38,10 @@ class BestFragment : BaseFragment<FragmentBestBinding>() {
         binding.rvBests.adapter = bestAdapter
         binding.rvBests.layoutManager = LinearLayoutManager(requireContext())
 
-        activityViewModel.getBests()
+        viewModel.getBests()
 
-        activityViewModel.bestItems
-            .flowWithLifecycle(this.lifecycle)
+        viewModel.bestItems
+            .flowWithLifecycle(viewLifecycleOwner.lifecycle)
             .onEach {
                 when (it) {
                     is UiState.Init -> {
