+++ app/src/main/java/co/kr/woowahan_banchan/presentation/viewmodel/main/MainDishViewModel.kt
@@ -7,7 +7,7 @@ import androidx.lifecycle.viewModelScope
 import co.kr.woowahan_banchan.domain.entity.dish.Dish
 import co.kr.woowahan_banchan.domain.repository.Source
 import co.kr.woowahan_banchan.domain.usecase.GetDishesUseCase
-import co.kr.woowahan_banchan.presentation.viewmodel.MainViewModel
+import co.kr.woowahan_banchan.presentation.viewmodel.UiState
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
@@ -48,17 +48,7 @@ class MainDishViewModel @Inject constructor(
         defaultMainDishes = list
     }
 
-    fun setModeGrid(){
-        _isGridMode.value = true
-    }
-
-    fun setModeLinear(){
-        _isGridMode.value = false
-    }
-
-    sealed class UiState<out T> {
-        object Init : UiState<Nothing>()
-        data class Success<out T>(val data: T) : UiState<T>()
-        data class Error(val message: String?) : UiState<Nothing>()
+    fun setGridMode(isGrid : Boolean){
+        _isGridMode.value = isGrid
     }
 }
