+++ app/src/main/res/layout/fragment_main_dish.xml
@@ -56,7 +56,7 @@
                     android:layout_width="24dp"
                     android:layout_height="24dp"
                     android:background="@drawable/selector_grid_tab"
-                    android:onClick="@{()->viewModel.setModeGrid()}"
+                    android:onClick="@{()->viewModel.setGridMode(true)}"
                     app:layout_constraintBottom_toBottomOf="parent"
                     app:layout_constraintStart_toStartOf="parent"
                     app:layout_constraintTop_toTopOf="parent"
@@ -67,7 +67,7 @@
                     android:layout_height="24dp"
                     android:layout_marginStart="4dp"
                     android:background="@drawable/selector_linear_tab"
-                    android:onClick="@{()->viewModel.setModeLinear()}"
+                    android:onClick="@{()->viewModel.setGridMode(false)}"
                     app:layout_constraintBottom_toBottomOf="parent"
                     app:layout_constraintStart_toEndOf="@id/ib_grid_btn"
                     app:layout_constraintTop_toTopOf="parent"
@@ -94,6 +94,7 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:layout_gravity="center_horizontal"
+            android:paddingBottom="30dp"
             app:layoutManager="androidx.recyclerview.widget.GridLayoutManager"
             app:layout_behavior="@string/appbar_scrolling_view_behavior"
             app:spanCount="2"
